move test_utils to test/utils.py
