+++ libfulltext/test/utils.py
